mallorca-real-estate-scraper/requirements.txt
googlemaps
pandas
python-dotenv
